pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "productivity-copilot-env"
version = "0.1.0"
description = "A Productivity Copilot OpenEnv Simulation"
authors = [
    {name = "productivity_hacker"}
]
dependencies = [
    "openenv-core",
    "pydantic",
    "numpy",
    "scikit-learn",
    "joblib",
    "sentence-transformers",
    "chromadb",
    "fastapi",
    "uvicorn",
    "openai",
    "huggingface_hub"
]

[project.optional-dependencies]
ml = []

[project.scripts]
server = "server.app:main"
